Add files via upload
Still need to change the constraints set for the longitude and latitude to be more appropriate (e.g. size of Europe? but also not too large-- i.e. where the amount of information is overwhelming). Add temperature graph, waiting for Federica's answer. Vertical velocity is in [1/s] rather than [m/s] for some reason, waiting for Federica's answer. Everything else good but needs to be analysed